+++ Application/ProfileService.Application.Services/FileAchievementServiceApp.cs
@@ -25,10 +25,10 @@ public class FileAchievementServiceApp : IFileAchievementServiceApp
     }
 
     /// <summary>
-    /// Получить .
+    /// Получить файл достижения.
     /// </summary>
-    /// <param name="id"> Идентификатор. </param>
-    /// <returns> ДТО . </returns>
+    /// <param name="id"> Идентификатор файла достижения. </param>
+    /// <returns> ДТО файла достижения. </returns>
     public async Task<IEnumerable<FileAchievementDto>> GetByAchievementIdAsync(int id)
     {
         List<FileAchievement> achievement = await _fileAchievementRepository.GetByAchievementIdAsync(id, CancellationToken.None);
@@ -37,54 +37,35 @@ public class FileAchievementServiceApp : IFileAchievementServiceApp
 
 
     /// <summary>
-    /// Создать .
+    /// Создать файл достижения.
     /// </summary>
-    /// <param name="creatingAchievementDto"> ДТО создаваемого . </param>
-    public async Task<int> CreateAsync(CreatingFileAchievementDto creatingFileAchievementDto)
+    /// <param name="achievementId"> id достижения. </param>
+    /// <param name="creatingFileAchievementDto"> ДТО создаваемого файла достижения. </param>
+    /// <returns>id файла достижения</returns>
+    public async Task<int> CreateAsync(int achievementId, CreatingFileAchievementDto creatingFileAchievementDto)
     {
-        var achievement = _mapper.Map<CreatingFileAchievementDto, FileAchievement>(creatingFileAchievementDto);
+        FileAchievement achievement = _mapper.Map<CreatingFileAchievementDto, FileAchievement>(creatingFileAchievementDto);
+        achievement.AchievementId = achievementId;
+
         var createdFileAchievement = await _fileAchievementRepository.AddAsync(achievement);
         await _fileAchievementRepository.SaveChangesAsync();
         return createdFileAchievement.Id;
     }
-    /// <summary>
-    /// Изменить .
-    /// </summary>
-    /// <param name="id"> Идентификатор. </param>
-    /// <param name="updatingFileAchievementDto"> ДТО редактируемого. </param>
-    //public async Task UpdateAsync(int id, UpdatingFileAchievementDto updatingFileAchievementDto)
-    //{
-    //    var fileAchievement = await _fileAchievementRepository.GetAsync(id, CancellationToken.None);
-    //    if (fileAchievement == null)
-    //    {
-    //        throw new Exception($"Профиль с идентфикатором {id} не найден");
-    //    }
-
-    //    fileAchievement.FileId = updatingFileAchievementDto.FileId;
-
-    //    _fileAchievementRepository.Update(fileAchievement);
-    //    await _fileAchievementRepository.SaveChangesAsync();
-    //}
 
     /// <summary>
-    /// Удалить .
+    /// Удалить файл достижения.
     /// </summary>
-    /// <param name="id"> Идентификатор . </param>
+    /// <param name="id"> Идентификатор файла достижения. </param>
+    /// <returns></returns>
     public async Task DeleteAsync(int id)
     {
         var fileAchievement = await _fileAchievementRepository.GetAsync(id, CancellationToken.None);
+        if (fileAchievement == null)
+        {
+            throw new Exception($"Файл достижения с идентфикатором {id} не найден");
+        }
+
         _fileAchievementRepository.Delete(fileAchievement);
         await _fileAchievementRepository.SaveChangesAsync();
     }
-    /// <summary>
-    /// Получить постраничный список.
-    /// </summary>
-    /// <param name="page"> Номер страницы. </param>
-    /// <param name="pageSize"> Объем страницы. </param>
-    /// <returns> Страница . </returns>
-    //public async Task<ICollection<FileAchievementDto>> GetPagedAsync(int page, int pageSize)
-    //{
-    //    ICollection<FileAchievement> entities = await _fileAchievementRepository.GetPagedAsync(page, pageSize);
-    //    return _mapper.Map<ICollection<FileAchievement>, ICollection<FileAchievementDto>>(entities);
-    //}
 }
